Update docker configuration and enable gunicorn
Update docker configuration to bring up the data node via docker-compose
so that each component lives in its own container rather than in a
single container.

Switch from Flask to gunicorn WSGI server; this necessitates ensuring
startup logic in storage_api.py is executed when main() is not used and
it is invalid to extract options from command line arguments.

Add an nginx reverse proxy to the data node to serve HTTPS, which allows
the SSL options to be removed from storage_api.py.

Add a Postman collection for testing of a local data node.

Fix a few typos and details, and get sync_token from header for issue #30.

Update all documentation as needed.