Record operation details from diagnostic logging tool (#81)
This change extends the TCL4 diagnostic `monitor` tool to capture operation details in addition to the grid summaries it previously captured. To do this, it creates a TokenManager to automatically retrieve and appropriately cache tokens with different scopes, and provides a differently-scoped access token to retrieve operations from USS's than to retrieve the grid state.

This tool was tested successfully during a UPP shakedown.