Fix Istio deployment + Db-manager errors
Istio no support Mesh Policy in v1alpha1 and there doesn't seem to be a replacement so its removed.

DB Manager job was failing to complete due to a bug in K8s when there is a side car injected. We don't need sidecar for this job and the current suggested work around is to remove it until K8s implements a fix