Add Index on end_at for improved performance (#445)
* Add Index on end_at for improved performance

Querying number of subscriptions is the bottle neck for inserting Subscriptions. Adding an index here making removes the full scan in the query. 
Ideally we want to index on cells however it is not supported in CRDB yet, They anticipate that v20.2 will support creating forward index on array

* Remove store Cells from index

Store cells cause transaction to fail with context deadline exceeded. It could be that Cockroach doesn't properly support forward indexing with arrays

* Update README

* Update README 2

* Update scripts to v3.1.1 or "latest" where ever possible